% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order-as.R
\name{align}
\alias{align}
\title{ordering of vector}
\usage{
align(x, template = NULL, group = NULL, all = TRUE, outgroup = ".Other")
}
\arguments{
\item{x}{the vector that needs ordering}

\item{template}{the wanted order}

\item{group}{possible grouping}

\item{all}{logical; should given elements not in template be kept?}

\item{outgroup}{group name of elements not in template (if kept)}
}
\description{
Order a given vector, which may contain duplicates, according to the wanted
  order given by some other vector. Supercedes 'order_as' and
  'order_by_list' in a single function.
}
