% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable-latex_format.R
\name{alt_num_format}
\alias{alt_num_format}
\title{alternative formating}
\usage{
alt_num_format(x, dg = 1, ds = 2, maybe.p = TRUE, p.bound = 0.001,
  miss = "", some.scientific = TRUE, low = 1e-08, high = 1e+08)
}
\arguments{
\item{x}{vector of numeric values}

\item{dg}{rounding digits for 'great' (abs>1) numbers}

\item{ds}{significance digits for 'small' numbers (abs<1)}

\item{maybe.p}{is x possible p-values?}

\item{p.bound}{if p-values, numbers below this will be '<p.bound'}

\item{miss}{character string to replace missing values}

\item{some.scientific}{possibly write some numbers in scientific notation}

\item{low}{lower threshold for when to resort to scientific notation}

\item{high}{upper threshold for when to resort to scientific notation}
}
\description{
alternative formating for numeric values, BUT IT DOES NOT SEEM TO DO WHAT I WANT
}
