% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-flist.R
\name{apply_flist}
\alias{apply_flist}
\title{apply flist}
\usage{
apply_flist(x, flist, ..., xname = NULL)
}
\arguments{
\item{x}{variable}

\item{flist}{list of functions}

\item{...}{arguments passed to the functions in flist}

\item{xname}{name of variable}
}
\description{
apply function list
}
\details{
apply the functions in flist to the variable 'x'
}

