% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable-misc.R
\name{as.data.frame.dtable}
\alias{as.data.frame.dtable}
\title{dtable to data.frame}
\usage{
\method{as.data.frame}{dtable}(x, ...)
}
\arguments{
\item{x}{dtable object}

\item{...}{arguments passed}
}
\description{
turn dtable into data.frame
}
\details{
this will just change the class attribute
}
