% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparers.R
\name{c_bnry}
\alias{c_bnry}
\alias{c_RR}
\alias{c_OR}
\alias{c_bstd}
\title{various comparer functions for bnry types}
\usage{
c_bnry(...)

c_RR(x, glist, w = NULL, ...)

c_OR(x, glist, w = NULL, ...)

c_bstd(x, glist, w = NULL, ...)
}
\arguments{
\item{...}{this is to be able to tolerate unnecessary arguments}

\item{x}{vector}

\item{glist}{a grouping list}

\item{w}{weight}
}
\description{
various comparer functions for bnry types
}
\section{Functions}{
\itemize{
\item \code{c_RR}: risk ratio, 2 groups only

\item \code{c_OR}: odds ratio, 2 groups only

\item \code{c_bstd}: standardized differences for binary variables, 2 groups only
}}

\references{
\url{https://www.lerner.ccf.org/qhs/software/lib/stddiff.pdf}
}
