% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flist-comparers.R
\name{c_date}
\alias{c_date}
\alias{c_overlap}
\title{various comparer functions for surv types}
\usage{
c_date(...)

c_overlap(x, glist, ...)
}
\arguments{
\item{...}{this is to be able to tolerate unnecessary arguments}

\item{x}{vector}

\item{glist}{a grouping list}
}
\description{
various comparer functions for surv types
}
\section{Functions}{
\itemize{
\item \code{c_overlap}: number of records in the overlapping time period, 2
groups only
}}

