% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flist-comparers.R
\name{c_surv}
\alias{c_rr}
\alias{c_surv}
\title{various comparer functions for surv types}
\usage{
c_surv(...)

c_rr(x, glist, w = NULL, type = "right", ...)
}
\arguments{
\item{...}{this is to be able to tolerate unnecessary arguments}

\item{x}{vector}

\item{glist}{a grouping list}

\item{w}{weight}

\item{type}{type of censoring}
}
\description{
various comparer functions for surv types
}
\section{Functions}{
\itemize{
\item \code{c_rr}: rate ratio, 2 groups only
}}

