% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-order-prune.R
\name{cbind_dtable}
\alias{cbind_dtable}
\title{rbind for dtables}
\usage{
cbind_dtable(x, y, groups = NULL)
}
\arguments{
\item{x}{object 1}

\item{y}{object 2}

\item{groups}{add meta info to the groups}
}
\description{
cbind dtables
}
\details{
wrapper for \cod{cbind} which keeps dtable attributes sane
}

