% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flist-describers.R
\name{d_bnry}
\alias{d_bn}
\alias{d_bnry}
\alias{d_bp}
\alias{d_odds}
\alias{d_ref_level}
\title{various describer functions for bnry types}
\usage{
d_bnry(...)

d_ref_level(x, ...)

d_bn(x, w = NULL, ...)

d_bp(x, w = NULL, ...)

d_odds(x, w = NULL, ...)
}
\arguments{
\item{...}{this is to be able to tolerate unnecessary arguments}

\item{x}{vector}

\item{w}{weight}
}
\description{
various describer functions for bnry types
}
\section{Functions}{
\itemize{
\item \code{d_ref_level}: return the reference level of a bnry factor

\item \code{d_bn}: number of occurences of the reference value

\item \code{d_bp}: proportion of occurences of the reference value

\item \code{d_odds}: the odds of the reference level
}}

