% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describers.R
\name{d_date}
\alias{d_date}
\alias{d_dmin}
\alias{d_dmax}
\title{various describing functions for date variables}
\usage{
d_date(...)

d_dmin(x, ...)

d_dmax(x, ...)
}
\arguments{
\item{...}{this is to be able to tolerate unnecessary arguments}

\item{x}{vector of dates}
}
\description{
various describing functions for date variables
}
\section{Functions}{
\itemize{
\item \code{d_dmin()}: minimum

\item \code{d_dmax()}: maximum

}}
