% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flist-describers.R
\name{d_real}
\alias{d_IQR}
\alias{d_length}
\alias{d_max}
\alias{d_mean}
\alias{d_median}
\alias{d_min}
\alias{d_missing}
\alias{d_missing.perc}
\alias{d_n}
\alias{d_real}
\alias{d_sd}
\alias{d_sum}
\title{various describer functions for real types}
\usage{
d_real(...)

d_n(x, w = NULL, ...)

d_length(x, ...)

d_missing(x, w = NULL, ...)

d_missing.perc(x, w = NULL, ...)

d_sum(x, w = NULL, ...)

d_mean(x, w = NULL, ...)

d_sd(x, w = NULL, ...)

d_median(x, w = NULL, ...)

d_min(x, ...)

d_max(x, ...)

d_IQR(x, ...)
}
\arguments{
\item{...}{this is to be able to tolerate unnecessary arguments}

\item{x}{vector}

\item{w}{weight}
}
\description{
various describer functions for real types
}
\section{Functions}{
\itemize{
\item \code{d_n}: number of non-missing elements

\item \code{d_length}: number of  elements

\item \code{d_missing}: Count missing

\item \code{d_missing.perc}: Percent missing

\item \code{d_sum}: sum of all elements

\item \code{d_mean}: mean

\item \code{d_sd}: standard deviation

\item \code{d_median}: median

\item \code{d_min}: minimum

\item \code{d_max}: maximum

\item \code{d_IQR}: inter quartile range
}}

