% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flist-describers.R
\name{d_surv}
\alias{d_dmax}
\alias{d_dmin}
\alias{d_esum}
\alias{d_rate}
\alias{d_surv}
\alias{d_tsum}
\title{various describing functions for surv variables}
\usage{
d_surv(...)

d_dmin(x, ...)

d_dmax(x, ...)

d_surv(...)

d_tsum(x, w = NULL, cens.type = "right", ...)

d_esum(x, w = NULL, cens.type = "right", ...)

d_rate(x, w = NULL, cens.type = "right", ...)
}
\arguments{
\item{...}{this is to be able to tolerate unnecessary arguments}

\item{x}{vector of dates}

\item{w}{weight}

\item{cens.type}{what kind of censoring?}

\item{x}{vector}

\item{...}{this is to be able to tolerate unnecessary arguments}
}
\description{
various describing functions for surv variables

various describer functions for surv types
}
\section{Functions}{
\itemize{
\item \code{d_dmin}: minimum

\item \code{d_dmax}: maximum

\item \code{d_tsum}: sum of follow up time

\item \code{d_esum}: sum of events

\item \code{d_rate}: rate (d_esum / d_tsum)
}}

