% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable-kable.R
\name{data_vlist2kbl}
\alias{data_vlist2kbl}
\title{data + vlist -> kable table (experimental)}
\usage{
data_vlist2kbl(
  data,
  guide = NULL,
  var.list = NULL,
  caption = NULL,
  label = NULL,
  ...,
  format = TRUE,
  format.param = as.list(NULL),
  n = c(n = "size"),
  tot.name = "All"
)
}
\arguments{
\item{data}{a data frame to be described}

\item{guide}{a dtable guide}

\item{var.list}{a variable list}

\item{caption}{caption}

\item{label}{label}

\item{...}{arguments passed to \code{dtables}}

\item{format}{logical; want formatting?}

\item{format.param}{list; formatting parameters}

\item{n}{size indicator in table (set to NULL to suppress this)}

\item{tot.name}{name of single column (if no glist)}
}
\description{
Create a dtables from data and make a nice table with "rgroup:s" using a
list that groups variables
}
