% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-set-attributes.R
\name{dattr<-}
\alias{dattr<-}
\title{change 'dtable' attribute}
\usage{
dattr(x) <- value
}
\arguments{
\item{x}{object}

\item{value}{'dtable' value}
}
\description{
get 'dtable'
}
\details{
shortcut and sanity check for \code{"attr<-"(x, "dtable", value)}
}
