% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-set-attributes.R
\name{dattr}
\alias{dattr}
\title{set 'dtable' attribute}
\usage{
dattr(x)
}
\arguments{
\item{x}{object}
}
\description{
shortcut and sanity check for \code{attr(x,
    "dtable")}
}
\details{
set 'dtable'
}
