% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delist.R
\name{delist}
\alias{delist}
\title{unlist alternative}
\usage{
delist(x)
}
\arguments{
\item{x}{a list}
}
\value{
vector
}
\description{
unlist a list of vectors, keeping vector names but not list names
}
\examples{
test <- list(foo = c("a" = 1), bar = c("b" = 2))
## unlist will generally (in my opinion) keep too many or too few names
unlist(test)
unlist(test, use.names = FALSE)
## delist will (for some purposes) do this better
delist(test)
## also, recursively
test2 <- list(baz = c("c" = 3), quuz = test)
delist(test2)
}
