% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{descripteur_other_types}
\alias{descripteur_other_types}
\title{other types}
\usage{
descripteur_other_types()
}
\description{
function that returns the types, other than describable ones, that can occur
    in a guide
}
