% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable-latex_format.R
\name{dformat_num}
\alias{dformat_num}
\title{numerical formating}
\usage{
dformat_num(
  x,
  dg = 1,
  ds = 2,
  maybe.p = TRUE,
  p.bound = 1e-04,
  miss = "",
  some.scientific = TRUE,
  low = 1e-08,
  high = 1e+08,
  ...,
  verbose = TRUE
)
}
\arguments{
\item{x}{vector of numeric values}

\item{dg}{rounding digits for 'great' (abs>1) numbers}

\item{ds}{significance digits for 'small' numbers (abs<1)}

\item{maybe.p}{logical; is x possibly p-values?}

\item{p.bound}{numeric; if p-values, numbers below this will be '<p.bound'}

\item{miss}{character string to replace missing values}

\item{some.scientific}{logical; possibly write some numbers in scientific notation?}

\item{low}{lower threshold for when to resort to scientific notation}

\item{high}{upper threshold for when to resort to scientific notation}

\item{...}{to be able to tolerate argument spamming}

\item{verbose}{logical; tell med things?}
}
\description{
numerical formating (mainly for \code{dtable_format})
}
