% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable-latex_format.R
\name{dformat_text}
\alias{dformat_text}
\title{text formating}
\usage{
dformat_text(x, tmax = 30, rep.key = NULL, add.key = TRUE, ..., verbose = TRUE)
}
\arguments{
\item{x}{vector of character values}

\item{tmax}{maximum number of chars}

\item{rep.key}{key of replacements, i.e. named vector of the form
'new value' = 'old value'}

\item{add.key}{logical; add given key.values to default ones?}

\item{...}{to be able to tolerate argument spamming}

\item{verbose}{logical; tell me things?}
}
\description{
text formating (mainly for \code{dtable_format})
}
