% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testers.R
\name{dt_comp}
\alias{dt_comp}
\alias{dt_empty_comp}
\alias{dt_empty_meta}
\alias{dt_wilcox.p}
\alias{dt_wilcox.p.info}
\alias{dt_kruskal.p}
\alias{dt_kruskal.p.info}
\alias{dt_fisher.p}
\alias{dt_fisher.p.info}
\alias{dt_chisq.p}
\alias{dt_chisq.p.info}
\title{various functions for compact summary of variables}
\usage{
dt_comp(...)

dt_empty_comp(x, ...)

dt_empty_meta(x, ...)

dt_wilcox.p(x, glist, ...)

dt_wilcox.p.info(x, glist, ...)

dt_kruskal.p(x, glist, ...)

dt_kruskal.p.info(x, glist, ...)

dt_fisher.p(x, glist, ...)

dt_fisher.p.info(x, glist, ...)

dt_chisq.p(x, glist, ...)

dt_chisq.p.info(x, glist, ...)
}
\arguments{
\item{...}{to be able to tolerate unnecessary arguments}

\item{x}{the data}

\item{glist}{grouping}
}
\description{
various functions for compact summary of variables
}
\section{Functions}{
\itemize{
\item \code{dt_empty_comp}: returns an empty string

\item \code{dt_empty_meta}: returns an empty string

\item \code{dt_wilcox.p}: wilcoxon test p-value

\item \code{dt_wilcox.p.info}: info on wilcoxon test p-value

\item \code{dt_kruskal.p}: kruskal-wallis test p-value

\item \code{dt_kruskal.p.info}: info on kruskal-wallis test p-value

\item \code{dt_fisher.p}: fishers exact test p-value

\item \code{dt_fisher.p.info}: info on fishers exact test p-value

\item \code{dt_chisq.p}: chisquare test p-value

\item \code{dt_chisq.p.info}: info on chisquare test p-value
}}

