% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describers.R
\name{dt_desc}
\alias{dt_desc}
\alias{dt_empty_desc}
\alias{dt_name}
\alias{dt_bname}
\alias{dt_cname}
\alias{dt_Q}
\alias{dt_Q.info}
\alias{dt_MedRange}
\alias{dt_MedRange.info}
\alias{dt_msd}
\alias{dt_msd.info}
\alias{dt_bcp}
\alias{dt_bcp.info}
\alias{dt_ccp}
\alias{dt_ccp.info}
\alias{dt_date}
\alias{dt_date.info}
\alias{dt_rate}
\alias{dt_rate.info}
\alias{dt_event}
\alias{dt_event.info}
\title{various functions for compact summary of variables}
\usage{
dt_desc(...)

dt_empty_desc(x, ...)

dt_name(x, xname = NULL, ...)

dt_bname(x, xname = NULL, ...)

dt_cname(x, xname = NULL, indent = "\\\\quad", ...)

dt_Q(x, useNA = FALSE, ...)

dt_Q.info(x, useNA = FALSE, ...)

dt_MedRange(x, useNA = FALSE, ...)

dt_MedRange.info(x, useNA = FALSE, ...)

dt_msd(x, useNA = FALSE, ...)

dt_msd.info(x, useNA = FALSE, ...)

dt_bcp(x, useNA = FALSE, ...)

dt_bcp.info(x, useNA = FALSE, ...)

dt_ccp(x, useNA = FALSE, ...)

dt_ccp.info(x, useNA = FALSE, ...)

dt_date(x, useNA = FALSE, ...)

dt_date.info(x, useNA = FALSE, ...)

dt_rate(x, useNA = FALSE, ...)

dt_rate.info(x, useNA = FALSE, ...)

dt_event(x, useNA = FALSE, ...)

dt_event.info(x, useNA = FALSE, ...)
}
\arguments{
\item{...}{arguments passed}

\item{x}{the data}

\item{xname}{name of variable}

\item{indent}{code for indentation}

\item{useNA}{display information on missing}
}
\description{
various functions for compact summary of variables
}
\section{Functions}{
\itemize{
\item \code{dt_empty_desc()}: returns an empty string

\item \code{dt_name()}: returns name

\item \code{dt_bname()}: returns name:ref for bnry

\item \code{dt_cname()}: returns name:ref for catg

\item \code{dt_Q()}: quantiles

\item \code{dt_Q.info()}: info for \code{dt_Q}

\item \code{dt_MedRange()}: med (min-max)

\item \code{dt_MedRange.info()}: info for \code{dt_MedRange}

\item \code{dt_msd()}: mean and standard deviation

\item \code{dt_msd.info()}: info for \code{dt_msd}

\item \code{dt_bcp()}: count and percentages (for bnry)

\item \code{dt_bcp.info()}: info for \code{dt_bcp}

\item \code{dt_ccp()}: count and percentages (for catg)

\item \code{dt_ccp.info()}: info for \code{dt_ccp}

\item \code{dt_date()}: first and last date

\item \code{dt_date.info()}: info for \code{dt_date}

\item \code{dt_rate()}: rate of events

\item \code{dt_rate.info()}: info for \code{dt_rate}

\item \code{dt_event()}: rate of events

\item \code{dt_event.info()}: info for \code{dt_rate}

}}
