% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flist-describers.R
\name{dt_fnc}
\alias{dt_Q}
\alias{dt_Q.info}
\alias{dt_bcp}
\alias{dt_bcp.info}
\alias{dt_bname}
\alias{dt_ccp}
\alias{dt_ccp.info}
\alias{dt_cname}
\alias{dt_date}
\alias{dt_date.info}
\alias{dt_empty_comp}
\alias{dt_empty_desc}
\alias{dt_event}
\alias{dt_event.info}
\alias{dt_fnc}
\alias{dt_msd}
\alias{dt_msd.info}
\alias{dt_name}
\alias{dt_rate}
\alias{dt_rate.info}
\title{various functions for compact summary of variables}
\usage{
dt_fnc(...)

dt_empty_desc(x, ...)

dt_empty_comp(x, ...)

dt_name(x, xname = NULL, ...)

dt_bname(x, xname = NULL, ...)

dt_cname(x, xname = NULL, ...)

dt_Q(x, useNA = FALSE, ...)

dt_Q.info(x, ...)

dt_msd(x, useNA = FALSE, ...)

dt_msd.info(x, ...)

dt_bcp(x, useNA = FALSE, ...)

dt_bcp.info(x, ...)

dt_ccp(x, useNA = FALSE, ...)

dt_ccp.info(x, ...)

dt_date(x, useNA = FALSE, ...)

dt_date.info(x, ...)

dt_rate(x, useNA = FALSE, ...)

dt_rate.info(x, ...)

dt_event(x, useNA = FALSE, ...)

dt_event.info(x, ...)
}
\arguments{
\item{...}{arguments passed}

\item{x}{the data}

\item{useNA}{display information on missing}
}
\description{
various functions for compact summary of variables
}
\section{Functions}{
\itemize{
\item \code{dt_empty_desc}: returns an empty string

\item \code{dt_empty_comp}: returns an empty string

\item \code{dt_name}: returns name

\item \code{dt_bname}: returns name:ref for bnry

\item \code{dt_cname}: returns name:ref for bnry

\item \code{dt_Q}: quantiles

\item \code{dt_Q.info}: info for \code{dt_Q}

\item \code{dt_msd}: mean and standard deviation

\item \code{dt_msd.info}: info for \code{dt_msd}

\item \code{dt_bcp}: count and percentages (for bnry)

\item \code{dt_bcp.info}: info for \code{dt_bcp}

\item \code{dt_ccp}: count and percentages (for catg)

\item \code{dt_ccp.info}: info for \code{dt_ccp}

\item \code{dt_date}: first and last date

\item \code{dt_date.info}: info for \code{dt_date}

\item \code{dt_rate}: rate of events

\item \code{dt_rate.info}: info for \code{dt_rate}

\item \code{dt_event}: rate of events

\item \code{dt_event.info}: info for \code{dt_rate}
}}

