% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable-attr.R
\name{dtable_attr}
\alias{dtable_attr}
\title{extract attributes}
\usage{
dtable_attr(dt, perc = FALSE, perc.sign = "\%")
}
\arguments{
\item{dt}{a dtable object}

\item{perc}{use percentages instead of counts}

\item{perc.sign}{what kind of percentage sign, if any}
}
\description{
extract tidy version of attributes from a dtable
}
