% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable-missing.R
\name{dtable_constants}
\alias{dtable_constants}
\title{get constants}
\usage{
dtable_constants(data, guide = NULL)
}
\arguments{
\item{data}{data frame or some such object}

\item{guide}{a guide (optional)}
}
\description{
get the value of the constants in a data set
}
