% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable-fnote.R
\name{dtable_fnote}
\alias{dtable_fnote}
\title{create footnotes from table variables}
\usage{
dtable_fnote(dt, text = NULL, footnote = NULL)
}
\arguments{
\item{dt}{a dtable}

\item{text}{the variables to add to 'info'}

\item{footnote}{the variables to add to 'info' and the variables to map upon,
or so it is thought to behave if ever implemented}
}
\description{
make \code{unique(variable)} a part of attribute 'info' and, if wanted,
    place a footnote on another variable
}

