% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable-latex.R
\name{dtable_format}
\alias{dtable_format}
\title{format a dtable}
\usage{
dtable_format(dt, param = as.list(NULL))
}
\arguments{
\item{dt}{a  dtable}

\item{param}{list of parameter values}
}
\description{
overall, low-precision formatting of dtable objects - quick and
    dirty way of getting something ok (hopefully), fast. could be
    developed... its weird argument structure (all parameters
    gathered in a list) is due to it being
    considered most useful when called from other functions.
}

