% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable-ignored.R
\name{dtable_ignored}
\alias{dtable_ignored}
\title{get ignored}
\usage{
dtable_ignored(data, guide = NULL)
}
\arguments{
\item{data}{data frame or some such object}

\item{guide}{a guide (optional)}
}
\description{
get the number of distinct data points for the ignored in a data set
}
