% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable-latex.R
\name{dtable_latex}
\alias{dtable_latex}
\title{latexify a dtable}
\usage{
dtable_latex(dt, bling = TRUE, format = FALSE, file = "", where = "htb",
  rowname = NULL, ..., format.param = as.list(NULL))
}
\arguments{
\item{dt}{a dtable}

\item{bling}{shall we compose suitable table bling from the
attributes? (default \code{TRUE})}

\item{format}{use \code{dtable_format}?}

\item{file}{(default empty string) passed to \code{Hmisc::latex}}

\item{where}{(default "htb") passed to \code{Hmisc::latex}}

\item{rowname}{(default \code{NULL}) passed to \code{Hmisc::latex}}

\item{...}{passed to \code{Hmisc::latex}}

\item{format.param}{list of parameters to pass to
\code{dtable_format} (only used if \code{format = TRUE}).}
}
\description{
process dtable into latex code via \code{Hmisc::latex}.
}

