% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable_missing.R
\name{dtable_missing}
\alias{dtable_missing}
\title{describe missing}
\usage{
dtable_missing(data = NULL, v = NULL, glist = NULL)
}
\arguments{
\item{data}{the data.frame or similar}

\item{v}{if \code{NULL} all variables are included, else a character vector of the
names of wanted variables or a formula (if data is null it will look for
the variables in the global workspace, but they need to be of the same length)}
}
\value{
a data.frame with
\itemize{
  \item{variable} name of variable
  \item{NA.n} number of \code{NA} in that variable
  \item{NA.perc} percent \code{NA} in that variable
  \item{n.assoc} number of variables with an OR > 1 and a p-value of <
    threshold p-value specified (according to fisher.test)
  \item{max.assoc} the variable name with the largest OR (>1) among those
    with a p < threshold p given
}
}
\description{
get some simple stats on missingness for a set of variables
}

