% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-order-prune.R
\name{dtable_order}
\alias{dtable_order}
\title{order dtable}
\usage{
dtable_order(x)
}
\arguments{
\item{x}{object}
}
\description{
order dtables according to meta info
}
\details{
place meta info to the left
}

