% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-order-prune.R
\name{dtable_prune}
\alias{dtable_prune}
\title{prune dtable}
\usage{
dtable_prune(x, rm = NULL, keep = NULL)
}
\arguments{
\item{x}{object}

\item{rm}{index or variable name to remove}

\item{keep}{index or variable name to keep}
}
\description{
prune dtable
}
\details{
remove columns by name or index
}

