% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable-std.R
\name{dtable_std}
\alias{dtable_std}
\title{Standardized differences}
\usage{
dtable_std(data = NULL, v = NULL, guide = NULL, glist,
  comp = "across", ...)
}
\arguments{
\item{data}{data frame or similar}

\item{v}{if \code{NULL} all variables are included, else a character vector of the
names of wanted variables or a formula (if data is null it will look for
the variables in the global workspace, but they need to be of the same length)}

\item{guide}{a guide (by \code{dtable_guide}), either to select variables OR
to provide unit information (the type-info will not be used)}

\item{glist}{an index list or name of grouping variable}

\item{comp}{(kind of) comparison}

\item{...}{arguments passed to \code{dtables}(\code{dtable})}
}
\description{
Table with standardized differences for all variable types
}
