% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable-bind_order_prune_subset.R
\name{dtable_subset}
\alias{dtable_subset}
\title{subset a dtable}
\usage{
dtable_subset(x, ..., all.attr = FALSE)
}
\arguments{
\item{x}{a dtable}

\item{...}{arguments passed to \code{subset}}

\item{all.attr}{keep more than just the essential attributes?}
}
\description{
select rows in a dtable
}
