% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable-unknown.R
\name{dtable_unknown}
\alias{dtable_unknown}
\title{get unknown}
\usage{
dtable_unknown(data, guide = NULL)
}
\arguments{
\item{data}{data frame or some such object}

\item{guide}{a guide (optional)}
}
\description{
get the class for the unknowns in a data set
}
