% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtables.R
\name{dtables}
\alias{dtables}
\title{dtable for multiple types}
\usage{
dtables(
  data,
  guide = NULL,
  ...,
  types = NULL,
  desc.flists = NULL,
  comp.flists = NULL,
  test.flists = NULL
)
}
\arguments{
\item{data}{the data set}

\item{guide}{a dtable guide}

\item{...}{arguments passed to dtable}

\item{types}{types wanted}

\item{desc.flists}{flists for description}

\item{comp.flists}{flists for comparison}

\item{test.flists}{flists for testing}
}
\description{
concatenate dtables for multiple types into a single dtable
}
