% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtables.R
\name{dtables2file_helper}
\alias{dtables2file_helper}
\alias{peek}
\title{for exporting dtables (experimental)}
\usage{
dtables2file_helper(dt, rm = NULL, reps = NULL, format = FALSE)

peek(dt)
}
\arguments{
\item{dt}{a dtables}

\item{rm}{columns to remove}

\item{reps}{vector of replacements, where the name replaces the entry,
e.g. \code{c('foo' = 'bar')} will replace all 'bar' with 'foo'}

\item{format}{use \code{dtable_format}?}
}
\description{
make a dtable ready to write to file
}
\section{Functions}{
\itemize{
\item \code{peek}: quick peek at dtable (in particular dtables)
}}

