% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtables.R
\name{dtables2latex}
\alias{dtables2latex}
\alias{dtables2latex_ungrouped_helper}
\alias{dtables2latex_grouped_helper}
\title{dtables to latex code}
\usage{
dtables2latex(
  dt,
  format = TRUE,
  format.param = as.list(NULL),
  n = c(n = "size"),
  ...
)

dtables2latex_ungrouped_helper(dt, format = TRUE, ...)

dtables2latex_grouped_helper(dt, format = TRUE, ...)
}
\arguments{
\item{dt}{a dtables}

\item{format}{logical; do you want to format?}

\item{format.param}{list of parameter values passed to format function}

\item{n}{size indicator in table (set to NULL to suppress this)}

\item{...}{arguments passed to \code{dtable_latex}}
}
\description{
dtables to latex code
}
\section{Functions}{
\itemize{
\item \code{dtables2latex_ungrouped_helper}: deprecated function

\item \code{dtables2latex_grouped_helper}: deprecated function
}}

