% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtables.R
\name{dtables2latex_ungrouped_helper}
\alias{dtables2latex_ungrouped_helper}
\alias{dtables2latex_grouped_helper}
\title{standard dtables to latex operations}
\usage{
dtables2latex_ungrouped_helper(dt, format = TRUE, ...)

dtables2latex_grouped_helper(dt, format = TRUE, ...)
}
\arguments{
\item{dt}{an object from \code{dtables}}

\item{format}{argument passed to \code{dtable_latex}}

\item{...}{arguments passed to \code{dtable_latex}}
}
\description{
helper function for common transformation of ungrouped dtables object into
    latex output; remove column 'variable' and put 'info' in info attribute
}
\section{Functions}{
\itemize{
\item \code{dtables2latex_grouped_helper}: helper function for common
transformation of grouped dtables object into; in addition to
\code{dtables2latex_ungrouped_helper}, also put column 'pinfo' as a
footnote to column 'p' (and format)
}}

