% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glist-functions.R
\name{factorize_glist}
\alias{factorize_glist}
\title{factorize a glist}
\usage{
factorize_glist(glist, as.factor = FALSE, reverse.levels = FALSE)
}
\arguments{
\item{glist}{a list of indices}

\item{as.factor}{return a factor object?}

\item{reverse.levels}{levels in order of glist?}
}
\description{
reverse-engineer a categorical variable from glist, if possible
}
