% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flist.R
\name{flist}
\alias{flist}
\title{flist}
\usage{
flist(x, dattr = NULL, local = FALSE)
}
\arguments{
\item{x}{vector of function names}

\item{dattr}{the dtable attribute}

\item{local}{look for function first in locally or 'descripteur'}
}
\description{
create function list
}
\details{
create a function list by specifying the names of functions,
    primarily searching in the descripteur namespace but also in
    other packages
}
