% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flist-functions.R
\name{flists}
\alias{flists}
\title{create flists}
\usage{
flists(real = FALSE, bnry = FALSE, catg = FALSE, date = FALSE,
  surv = FALSE, names = NULL, thing = "desc")
}
\arguments{
\item{real}{logical or an flist}

\item{bnry}{logical or an flist}

\item{catg}{logical or an flist}

\item{date}{logical or an flist}

\item{surv}{logical or an flist}

\item{names}{row names wanted}

\item{thing}{character; 'desc', 'comp' or 'test'}
}
\description{
flists is a list of one or more flist, one for each type wanted (and named
thereafter)
}
