% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable-guide.r
\name{get_class}
\alias{get_class}
\title{get class}
\usage{
get_class(x)
}
\arguments{
\item{x}{a vector}
}
\description{
get/infer the class of a vector, giving priority to numeric, integer,
    character, factor, logical, Surv, Date, POSIXct and POSIXt
}
