% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delist.R
\name{invert_list}
\alias{invert_list}
\title{invert list of named characters}
\usage{
invert_list(x)
}
\arguments{
\item{x}{a list of named vectors}
}
\description{
switch names/values in a list of named vectors
}
\examples{
test <- list(foo = c("a" = "first letter"),
             bar = c("b" = "second letter",
                     "c" = "third letter"))
invert_list(test)
}
