% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex-functions.R
\name{latex_figure}
\alias{latex_figure}
\title{LaTeX code for a figure}
\usage{
latex_figure(
  path,
  caption = NULL,
  label = NULL,
  cat = TRUE,
  lab.prefix = "fig:",
  where = "htb"
)
}
\arguments{
\item{path}{path to figure}

\item{caption}{caption of figure}

\item{label}{label of figure}

\item{cat}{concatenate and print the code, else return as character}

\item{lab.prefix}{prefix for label}

\item{where}{argument for placement of figure}
}
\value{
printed coded or character
}
\description{
generate the LaTeX code needed to produce a figure, with caption and label. A
    caption and label will be generated if unspecified
}
