% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-order-prune.R
\name{meta_order_dtable}
\alias{meta_order_dtable}
\title{order dtable}
\usage{
meta_order_dtable(x)
}
\arguments{
\item{x}{object}
}
\description{
order dtables according to meta info
}
\details{
place meta info to the left
}

