% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{opts_desc}
\alias{opts_desc}
\title{desc options}
\format{An object of class \code{list} of length 4.}
\usage{
opts_desc
}
\description{
This list tries to mimic the behavior of opts_chunk from knitr.
Currently these values are maintained with the functions in (the list)
\code{opts_desc}:
\itemize{
\item value - default: foo
}
}
\keyword{datasets}
