% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order-as.R
\name{order_as}
\alias{order_as}
\title{ordering of vector}
\usage{
order_as(given, wanted, incl.unordered = TRUE)
}
\arguments{
\item{given}{the vector that needs ordering}

\item{wanted}{the wanted order}

\item{incl.unordered}{should given elements not in wanted be kept?}
}
\value{
an index vector
}
\description{
order a given vector, which may contain duplicates, according to the
  wanted order given by some other vector
}
\examples{
g <- c("b", "d", "b", "f", "c", "c", "a")
order_as(given = g, wanted = letters[1:4])
g[order_as(given = g, wanted = letters[1:4])]
g[order_as(given = g, wanted = letters[1:4], incl.unordered = FALSE)]
}
