% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order-as.R
\name{order_by_list}
\alias{order_by_list}
\alias{order_as_list}
\title{ordering of vector by list}
\usage{
order_by_list(
  given,
  wanted,
  incl.unordered = TRUE,
  unordered.label = "Unlabelled"
)

order_as_list(
  given,
  wanted,
  incl.unordered = TRUE,
  unordered.label = "Unlabelled"
)
}
\arguments{
\item{given}{the vector that needs ordering}

\item{wanted}{the wanted order (list)}

\item{incl.unordered}{should given elements not in wanted be kept?}

\item{unordered.label}{label for the not-in-wanted elements}
}
\value{
list consisting of an index vector and a dataframe that connects the
  sorted elements with the corresponding list names. Also, the elements of
  \code{rle} applied to list.names entry of the dataframe
}
\description{
order a given vector, which may contain duplicates, according to the
  wanted order given by some list
}
\examples{
g <- c("b", "d", "b", "f", "c", "c", "a")
w <- list(
    "FOO" = c("a", "b"),
    "BAR" = c("c"),
    "BAZ" = c("e"),
    "QUUZ" = c("d")
)
order_by_list(g, w)
order_by_list(g, w, incl.unordered = FALSE)
}
