% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable-print.R
\name{print.dtable}
\alias{print.dtable}
\title{print dtable}
\usage{
\method{print}{dtable}(x, ...)
}
\arguments{
\item{x}{dtable object}

\item{...}{arguments passed}
}
\description{
print dtable object
}
\details{
prints the data.frame part as well as the dtable attributes
    if they are sane
}

