% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable_guide.r
\name{print.dtable_guide}
\alias{print.dtable_guide}
\title{print dtable_guide objects}
\usage{
\method{print}{dtable_guide}(x, ...)
}
\arguments{
\item{x}{a \code{dtable_guide } object}

\item{...}{arguments passed to \code{\link[base]{print.data.frame}}}
}
\description{
prints the data source attribute as well as the data frame
}

