% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable-other.R
\name{print.dtable_other}
\alias{print.dtable_other}
\title{print method for dtable_other objects}
\usage{
\method{print}{dtable_other}(x, ..., sep = "\\n")
}
\arguments{
\item{x}{the object}

\item{...}{for compatability}

\item{sep}{if multiple dtable_other attributes, how to separate them}
}
\description{
dtable_other objects are objects with an additional attribute 'dtable_other'
    which will be printed alongside the inherited print method
}
