% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-order-prune.R
\name{prune_dtable}
\alias{prune_dtable}
\title{prune dtable}
\usage{
prune_dtable(x, rm = NULL)
}
\arguments{
\item{x}{object}

\item{rm}{index or variable name to remove}
}
\description{
prune dtable
}
\details{
remove columns by name or index
}

