% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-order-prune.R
\name{rbind_dtable}
\alias{rbind_dtable}
\title{rbind for dtables}
\usage{
rbind_dtable(x, y)
}
\arguments{
\item{x}{object 1}

\item{y}{object 2}
}
\description{
rbind dtables
}
\details{
wrapper for \code{rbind}
}

