% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testers.R
\name{t_bnry}
\alias{t_bnry}
\alias{t_fisher.p}
\alias{t_chisq.p}
\title{various testing functions for bnry types}
\usage{
t_bnry(...)

t_fisher.p(x, glist, ...)

t_chisq.p(x, glist, ...)
}
\arguments{
\item{...}{this is to be able to tolerate unnecessary arguments}

\item{x}{vector}

\item{glist}{a grouping list}

\item{w}{weight}
}
\description{
various testing functions for bnry types
}
\section{Functions}{
\itemize{
\item \code{t_fisher.p}: fisher test p-value

\item \code{t_chisq.p}: chisquare test p-value
}}

