% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testers.R
\name{t_catg}
\alias{t_catg}
\alias{t_cchisq.p}
\title{various testing functions for catg types}
\usage{
t_catg(...)

t_cchisq.p(x, glist, ...)
}
\arguments{
\item{...}{this is to be able to tolerate unnecessary arguments}

\item{x}{vector}

\item{glist}{a grouping list}
}
\description{
various testing functions for catg types
}
\section{Functions}{
\itemize{
\item \code{t_cchisq.p()}: chisquare test p-value

}}
\seealso{
\code{\link{t_fisher.p}} and \code{\link{t_chisq.p}} which work for
    catg variables as well
}
