% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testers.R
\name{t_date}
\alias{t_date}
\alias{t_date_empty}
\title{various testing functions for date types}
\usage{
t_date(...)

t_date_empty(x, glist, ...)
}
\arguments{
\item{...}{this is to be able to tolerate unnecessary arguments}

\item{x}{vector}

\item{glist}{a grouping list}
}
\description{
various testing functions for date types
}
\section{Functions}{
\itemize{
\item \code{t_date_empty}: empty test
}}

