% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testers.R
\name{t_real}
\alias{t_real}
\alias{t_t.test.p}
\alias{t_wilcox.p}
\alias{t_kruskal.p}
\alias{t_anova.p}
\title{various testing functions for real types}
\usage{
t_real(...)

t_t.test.p(x, glist, ...)

t_wilcox.p(x, glist, ...)

t_kruskal.p(x, glist, ...)

t_anova.p(x, glist, ...)
}
\arguments{
\item{...}{this is to be able to tolerate unnecessary arguments}

\item{x}{vector}

\item{glist}{a grouping list}

\item{w}{weight}
}
\description{
various testing functions for real types
}
\section{Functions}{
\itemize{
\item \code{t_t.test.p}: t.test p-value, 2 groups only (could be made with weights
but not implemented yet)

\item \code{t_wilcox.p}: wilcoxon test p.value

\item \code{t_kruskal.p}: kruskal-wallis p-value, any number of groups

\item \code{t_anova.p}: ANOVA test p-value, any number of groups
}}

