% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testers.R
\name{t_surv}
\alias{t_surv}
\alias{t_surv_empty}
\title{various testing functions for surv types}
\usage{
t_surv(...)

t_surv_empty(x, glist, ...)
}
\arguments{
\item{...}{this is to be able to tolerate unnecessary arguments}

\item{x}{vector}

\item{glist}{a grouping list}

\item{w}{weight}

\item{cens.type}{type of censoring}
}
\description{
various testing functions for surv types
}
\section{Functions}{
\itemize{
\item \code{t_surv_empty}: empty test
}}

