% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{translate}
\usage{
translate(s, key, flexible = TRUE)
}
\arguments{
\item{s}{vector to translate}

\item{key}{translation key}

\item{flexible}{allow a key to be inversed if that seems more plausible?}
}
\value{
vector
}
\description{
translate by key
}
\details{
if you  have a key \code{key <- c('a'  = 'A', 'b' = 'B')} you  can use it to
    translate \code{s  <- c('a', 'b',  'a')} to \code{c('A', 'B',  'A')} via
    \code{key[v]}, which is  nice. But if \code{s <- c('a',  'b', 'c')} then
    \code{key[s]} is  \code{c('A', 'B', NA)}. The  function \code{translate}
    will  make sure  that translations  occur whenever  a corresponding  key
    entry exists, else keep the orginal value.
}
