% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describers.R
\name{weighted.quantile}
\alias{weighted.quantile}
\title{weighted quantile}
\usage{
weighted.quantile(x, probs = seq(0, 1, 0.25), w = NULL)
}
\arguments{
\item{x}{numeric vector whose sample quantiles are wanted}

\item{probs}{numeric vector of probabilities with values in [0,1]}

\item{w}{a numerical vector of weights the same length as 'x' giving
the weights to use for elements of 'x'}
}
\description{
Compute weighted quantiles
}
