% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr2text.R
\name{attr2text}
\alias{attr2text}
\title{get attributes as text}
\usage{
attr2text(dt, perc = FALSE, perc.sign = "\%", lessthan = "<",
  attr = c("size", "cc", "weight", "units", "info"), sep = ". ",
  vector = FALSE, rm.if.all = FALSE)
}
\arguments{
\item{dt}{a dtable}

\item{perc}{use percentages}

\item{perc.sign}{percentage sign}

\item{lessthan}{'strictly less than' sign}

\item{attr}{which attributes to extract}

\item{sep}{separator between attributes for output}

\item{vector}{return vector instead of a single string}

\item{rm.if.all}{exclude info on units if all unique}
}
\description{
extract attributes of dtable to a string (or character vector)
}
