% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparers.R
\name{c_date}
\alias{c_date}
\alias{c_overlap}
\alias{c_dstd}
\title{various comparer functions for date types}
\usage{
c_date(...)

c_overlap(x, glist, w = NULL, ...)

c_dstd(x, glist, ...)
}
\arguments{
\item{...}{this is to be able to tolerate unnecessary arguments}

\item{x}{vector}

\item{glist}{a grouping list}

\item{w}{weights}
}
\description{
various comparer functions for date types
}
\section{Functions}{
\itemize{
\item \code{c_overlap}: proportion of records in the overlapping time period, 2
groups only

\item \code{c_dstd}: standardized difference for date variables (same as for
real variables when the dates are interpreted as numericals)
}}

