% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparers.R
\name{c_real}
\alias{c_real}
\alias{c_shift}
\alias{c_rstd}
\title{various comparer functions for real types}
\usage{
c_real(...)

c_shift(x, glist, ...)

c_rstd(x, glist, w = NULL, ...)
}
\arguments{
\item{...}{this is to be able to tolerate unnecessary arguments}

\item{x}{vector}

\item{glist}{a grouping list}

\item{w}{weight}
}
\description{
various comparer functions for real types
}
\section{Functions}{
\itemize{
\item \code{c_shift}: median shift measured in units of original IQR, 2 groups only

\item \code{c_rstd}: standardized difference (2 groups only) assuming
independence
}}

\references{
\url{https://www.lerner.ccf.org/qhs/software/lib/stddiff.pdf}
}
