% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparers.R
\name{c_surv}
\alias{c_surv}
\alias{c_rr}
\alias{c_sstd}
\alias{c_cph}
\title{various comparer functions for surv types}
\usage{
c_surv(...)

c_rr(x, glist, w = NULL, cens.type = "right", ...)

c_sstd(x, glist, w = NULL, cens.type = "right", ...)

c_cph(x, glist, w = NULL, cens.type = "right", ...)
}
\arguments{
\item{...}{this is to be able to tolerate unnecessary arguments}

\item{x}{vector}

\item{glist}{a grouping list}

\item{w}{weight}

\item{cens.type}{type of censoring}
}
\description{
various comparer functions for surv types
}
\section{Functions}{
\itemize{
\item \code{c_rr}: rate ratio, 2 groups only

\item \code{c_sstd}: standardized difference between the rates of two surv
variables (difference in mean divided by 'average sd'-isch)

\item \code{c_cph}: Hazard Ratio from Cox model
}}

