% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flist-describers.R
\name{d_date}
\alias{d_date}
\alias{d_dmax}
\alias{d_dmin}
\title{various describing functions for surv variables}
\usage{
d_date(...)

d_dmin(x, ...)

d_dmax(x, ...)
}
\arguments{
\item{...}{this is to be able to tolerate unnecessary arguments}

\item{x}{vector of dates}
}
\description{
various describing functions for surv variables
}
\section{Functions}{
\itemize{
\item \code{d_dmin}: minimum

\item \code{d_dmax}: maximum
}}

