% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flist-describers.R
\name{d_surv}
\alias{d_esum}
\alias{d_rate}
\alias{d_surv}
\alias{d_tsum}
\title{various describer functions for surv types}
\usage{
d_surv(...)

d_tsum(x, w = NULL, type = "right", ...)

d_esum(x, w = NULL, type = "right", ...)

d_rate(x, w = NULL, type = "right", ...)
}
\arguments{
\item{...}{this is to be able to tolerate unnecessary arguments}

\item{x}{vector}

\item{w}{weight}

\item{type}{what kind of censoring?}
}
\description{
various describer functions for surv types
}
\section{Functions}{
\itemize{
\item \code{d_tsum}: sum of follow up time

\item \code{d_esum}: sum of events

\item \code{d_rate}: rate (d_esum / d_tsum)
}}

