% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describers.R
\name{d_surv}
\alias{d_surv}
\alias{d_tsum}
\alias{d_esum}
\alias{d_rate}
\title{various describer functions for surv types}
\usage{
d_surv(...)

d_tsum(x, w = NULL, weight = "sample", cens.type = "right", ...)

d_esum(x, w = NULL, weight = "sample", cens.type = "right", ...)

d_rate(x, w = NULL, cens.type = "right", ...)
}
\arguments{
\item{...}{this is to be able to tolerate unnecessary arguments}

\item{x}{vector}

\item{w}{weight}

\item{weight}{style of weighting, 'sample' (default) or 'case'}

\item{cens.type}{what kind of censoring?}
}
\description{
various describer functions for surv types
}
\section{Functions}{
\itemize{
\item \code{d_tsum()}: sum of follow up time

\item \code{d_esum()}: sum of events

\item \code{d_rate()}: rate (d_esum / d_tsum)

}}
