% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable.R
\name{dtable}
\alias{dtable}
\title{create descriptive table}
\usage{
dtable(data, type, guide = NULL, desc = TRUE, comp = NULL, glist = NULL,
  w = NULL, useNA = "ifany", ...)
}
\arguments{
\item{data}{a \code{data.frame} or such object}

\item{type}{what kind of variables to you want to describe?}

\item{guide}{point to guide or get the default}

\item{desc}{want to decribe variables?}

\item{comp}{want to apply a flist of comparers? (requires a glist)}

\item{glist}{grouping list, if wanted}

\item{w}{weights}

\item{useNA}{how to handle \code{NA} (requires that functions in
flist has this argument)}

\item{...}{arguments passed}
}
\description{
create descriptive table
}
\details{
create a description of variables of a given type in your
    data. this functions should have a vignette...
}

