% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable-fnote.R
\name{dtable_fnote}
\alias{dtable_fnote}
\title{create (latex) footnotes from table variables}
\usage{
dtable_fnote(
  dt,
  info,
  fn.var,
  info.attr = "info",
  format = FALSE,
  format.param = as.list(NULL)
)
}
\arguments{
\item{dt}{a dtable}

\item{info}{variable with the footnote info}

\item{fn.var}{variable to get footnotes}

\item{info.attr}{name of attribute to store info in}

\item{format}{format the dtable?}

\item{format.param}{parameters for formatting}
}
\description{
make \code{unique(variable)} a part of attribute 'info' and
    place a footnote on another variable
}
