% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable-latex.R
\name{dtable_format}
\alias{dtable_format}
\title{format a dtable}
\usage{
dtable_format(dt, b = 1, bh = 1, hfnc = base::round, bl = 2,
  lfnc = base::signif, br = 2, rfnc = base::round, p_b = 1e-04,
  peq0 = TRUE, tmax = 30, repus = TRUE)
}
\arguments{
\item{dt}{a dtable}

\item{b}{boundary, if numbers are consistenly higher they are handled by
'bh' and 'hfnc', if consistently lower by 'bl' and 'lfnc', and otherwise
by 'br' and 'rfnc'}

\item{bh}{digits argument for \code{hfnc}}

\item{hfnc}{format function for above boundary numbers}

\item{bl}{digits argument for \code{lfnc}}

\item{lfnc}{format function for below boundary numbers}

\item{br}{digits argument for \code{rfnc}}

\item{rfnc}{format function for other numbers}

\item{p_b}{threshold for how small p values to show. Any variable which is
between 0 and 1 is considered a 'p-value' here, which need not be the
case.}

\item{peq0}{even if we abbreviate small 'p-values' should we explicitly put
= "0" if it is equal to zero?}

\item{tmax}{how many characters to print for a character vector}

\item{repus}{should we replace "_" with "\\_" in charcter variables? (If not
LaTeX might fail.)}
}
\description{
overall, low-precision formatting of dtable objects - quick and
    dirty way of getting something ok (hopefully), fast. could be
    developed...
}
