% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable-latex_format.R
\name{dtable_format}
\alias{dtable_format}
\title{formatting}
\usage{
dtable_format(dt, ...)
}
\arguments{
\item{dt}{a dtable or such}

\item{...}{arguments passed to \code{dformat_num} and \code{dformat_text}}
}
\description{
formatting of dtables and data.frames
}
