% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable_guide.r
\name{dtable_guide}
\alias{dtable_guide}
\title{Description guide}
\usage{
dtable_guide(data, elim.set = NULL, catg.tol = 6, real.tol = catg.tol,
  date.tol = catg.tol, as.is = FALSE)
}
\arguments{
\item{data}{a data frame}

\item{elim.set}{names of variables to exclude from description}

\item{catg.tol}{only describe categorical data with no more than this
many unique values}

\item{real.tol}{force numeric data with few ($<= \code{real.tol}) unique data
points to be describes as categorical}

\item{date.tol}{force date data with few ($<= \code{date.tol}) unique data
points to be describes as categorical}

\item{as.is}{if TRUE ignore all tolerence parameters}
}
\description{
Determine how to describe a data frame
}

