% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable-latex_format.R, R/dtable-list_latex.R
\name{dtable_latex}
\alias{dtable_latex}
\alias{dtable_list_latex}
\title{latexify a dtable}
\usage{
dtable_latex(dt, bling = TRUE, bling.param = as.list(NULL), file = "",
  where = "htb", rowname = NULL, grey = NULL, ..., no.below = FALSE,
  guide = NULL, format = FALSE, format.param = as.list(NULL))

dtable_list_latex(dtl, ..., label, caption, use.sprintf = TRUE)
}
\arguments{
\item{dt}{a dtable}

\item{bling}{shall we compose suitable table bling from the
attributes? (default \code{TRUE})}

\item{bling.param}{list of parameters sent to \code{attr2text}, if non-empty \code{bling}
will be set to TRUE}

\item{file}{(default empty string) passed to \code{Hmisc::latex}}

\item{where}{(default "htb") passed to \code{Hmisc::latex}}

\item{rowname}{(default \code{NULL}) passed to \code{Hmisc::latex}}

\item{grey}{should some lines be gray to improve readability? This
requires that '\\usepackage[table]\{xcolor\}' be added to the
preamble. If \code{TRUE} then every other line is grey, but this can
also be specifed as a variable name or a given vector, see the vignette
'describe-data' for details.}

\item{...}{passed to \code{Hmisc::latex}}

\item{no.below}{shortcut to remove the bling that is places below the table
(hence will only do something if \code{bling = TRUE})}

\item{guide}{a guide to provide labels, minimally a data frame with
variables 'variable' and 'label'}

\item{format}{use \code{dtable_format}?}

\item{format.param}{list of parameters to pass to
\code{dtable_format} (if given will set \code{format = TRUE}).}

\item{dtl}{a list of dtables}

\item{label}{labels for the tables}

\item{caption}{captions for the tables}

\item{use.sprintf}{if TRUE then labels and captions should be uniform in
text apart from the specification of type, where a '\%s' should be.}
}
\description{
process dtable into latex code via \code{Hmisc::latex}
}
\section{Functions}{
\itemize{
\item \code{dtable_list_latex}: create multiple latex tables from a dtable list
}}

