% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable-latex.R, R/dtable-list.R
\name{dtable_latex}
\alias{dtable_latex}
\alias{dtable_list_latex}
\title{latexify a dtable}
\usage{
dtable_latex(dt, bling = TRUE, bling.param = as.list(NULL), file = "",
  where = "htb", rowname = NULL, ..., guide = NULL, format = FALSE,
  format.param = as.list(NULL))

dtable_list_latex(dtl, ..., label, caption, use.sprintf = TRUE)
}
\arguments{
\item{dt}{a dtable}

\item{bling}{shall we compose suitable table bling from the
attributes? (default \code{TRUE})}

\item{bling.param}{list of parameters sent to attr2text, if non-empty bling
will be set to TRUE}

\item{file}{(default empty string) passed to \code{Hmisc::latex}}

\item{where}{(default "htb") passed to \code{Hmisc::latex}}

\item{rowname}{(default \code{NULL}) passed to \code{Hmisc::latex}}

\item{...}{passed to \code{Hmisc::latex}}

\item{guide}{a guide to provide labels, minimally a data frame with
variables 'variable' and 'label'}

\item{format}{use \code{dtable_format}?}

\item{format.param}{list of parameters to pass to
\code{dtable_format} (if given will set \code{format = TRUE}).}

\item{dtl}{a list of dtables}

\item{label}{labels for the tables}

\item{caption}{captions for the tables}

\item{use.sprintf}{if TRUE then labels and captions should be uniform in
text apart from the specification of type, where a '%s' should be.}

\item{...}{arguments passed to \code{dtable_latex}}
}
\description{
process dtable into latex code via \code{Hmisc::latex}.
}
\section{Functions}{
\itemize{
\item \code{dtable_list_latex}: create multiple latex tables from a dtable list
}}

