% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable_missing.R
\name{dtable_missing}
\alias{dtable_missing}
\title{describe missing}
\usage{
dtable_missing(data = NULL, v = NULL, guide = NULL, glist = NULL)
}
\arguments{
\item{data}{the data.frame or similar}

\item{v}{if \code{NULL} all variables are included, else a character vector of the
names of wanted variables or a formula (if data is null it will look for
the variables in the global workspace, but they need to be of the same length)}

\item{guide}{a guide (by \code{dtable_guide}), either to select variables OR
to provide unit information (the type-info will not be used)}

\item{glist}{an index list or name of grouping variable}
}
\value{
a data.frame with
\itemize{
  \item{variable} name of variable
  \item{count} number of \code{NA} in that variable
  \item{percent} percent \code{NA} in that variable
}
}
\description{
get some simple stats on missingness for a set of variables
}

