% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable-bind_order_prune_subset.R
\name{dtable_prune}
\alias{dtable_prune}
\title{prune dtable}
\usage{
dtable_prune(
  x,
  rm = NULL,
  keep = NULL,
  info = FALSE,
  info.attr = "info",
  info.unique = TRUE,
  split.unique = TRUE
)
}
\arguments{
\item{x}{object}

\item{rm}{index or variable name to remove}

\item{keep}{index or variable name to keep (specify this or 'rm' and not both)}

\item{info}{store the discarded information in attributes?}

\item{info.attr}{name of attribute to store discarded info (if \code{info = TRUE})}

\item{info.unique}{store only unique info (if \code{info = TRUE})}

\item{split.unique}{if \code{unique.info = TRUE}, also split into individual
sentences before determining uniqueness?}
}
\description{
prune dtable
}
\details{
remove columns by name or index
}
