% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtable.R
\name{dtables}
\alias{dtables}
\title{dtable for multiple types}
\usage{
dtables(data, types = NULL, desc.flists = NULL, comp.flists = NULL,
  guide = NULL, ...)
}
\arguments{
\item{data}{the data set}

\item{types}{types wanted}

\item{desc.flists}{flists for description}

\item{comp.flists}{flists for comparison}

\item{guide}{a dtable guide}

\item{...}{arguments passed to dtable}
}
\description{
concatenate dtables for mutiple types into a single dtable
}

