% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make--factorize-glist.R
\name{factorize_glist}
\alias{factorize_glist}
\title{factorize a glist}
\usage{
factorize_glist(glist)
}
\arguments{
\item{glist}{a glist (duh!)}
}
\description{
reverse-engineer a categorical variable from glist, if possible
}

