% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-glist.R
\name{make_glist}
\alias{make_glist}
\title{make glist}
\usage{
make_glist(x, ref = NULL)
}
\arguments{
\item{x}{object}

\item{ref}{reference}
}
\description{
make a grouping list
}
\details{
description tables with grouping requires a list of indexes to
    group on. Often this should be equivalent to categorization
    according to some grouping variable. If so, the 'glist' needed
    can be created by either specifying that variable or the name
    of that variable in data source 'ref'.
}

