% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testers.R
\name{t_surv}
\alias{t_surv}
\alias{t_surv_empty}
\alias{t_HR.p}
\alias{t_lr.p}
\title{various testing functions for surv types}
\usage{
t_surv(...)

t_surv_empty(x, glist, ...)

t_HR.p(x, glist, w = NULL, cens.type = "right", ...)

t_lr.p(x, glist, w = NULL, cens.type = "right", ...)
}
\arguments{
\item{...}{this is to be able to tolerate unnecessary arguments}

\item{x}{vector}

\item{glist}{a grouping list}

\item{w}{weight}

\item{cens.type}{type of censuring}
}
\description{
various testing functions for surv types
}
\section{Functions}{
\itemize{
\item \code{t_surv_empty()}: empty test

\item \code{t_HR.p()}: p-value from Cox model

\item \code{t_lr.p()}: p-value from log-rank test

}}
